added userdto and updated

+++ Entities/Concretes/Rank.cs
@@ -11,7 +11,7 @@ namespace Entities.Concretes
     {
         public int Id { get; set; }
         public string Name { get; set; }
-        public int Level { get; set; }
+        public short Level { get; set; }
         public string Color { get; set; }
     }
 }
